+++ packages/headless/src/features/question-answering/question-answering-analytics-actions.ts
@@ -12,6 +12,11 @@ import {
   QuestionAnsweringDocumentIdActionCreatorPayload,
 } from './question-answering-document-id';
 
+export type SmartSnippetFeedback =
+  | 'does_not_answer'
+  | 'partially_answers'
+  | 'was_not_a_question';
+
 export const logExpandSmartSnippet = makeAnalyticsAction(
   'analytics/smartSnippet/expand',
   AnalyticsType.Custom,
@@ -47,6 +52,32 @@ export const logOpenSmartSnippetSource = (source: Result) =>
         documentIdentifier(source)
       );
     }
+  )();
+
+export const logOpenSmartSnippetFeedbackModal = makeAnalyticsAction(
+  'analytics/smartSnippet/feedbackModal/open',
+  AnalyticsType.Custom,
+  (client) => client.logOpenSmartSnippetFeedbackModal()
+);
+
+export const logCloseSmartSnippetFeedbackModal = makeAnalyticsAction(
+  'analytics/smartSnippet/feedbackModal/close',
+  AnalyticsType.Custom,
+  (client) => client.logCloseSmartSnippetFeedbackModal()
+);
+
+export const logSmartSnippetFeedback = (feedback: SmartSnippetFeedback) =>
+  makeAnalyticsAction(
+    'analytics/smartSnippet/sendFeedback',
+    AnalyticsType.Custom,
+    (client) => client.logSmartSnippetFeedbackReason(feedback)
+  )();
+
+export const logSmartSnippetDetailedFeedback = (details: string) =>
+  makeAnalyticsAction(
+    'analytics/smartSnippet/sendFeedback',
+    AnalyticsType.Custom,
+    (client) => client.logSmartSnippetFeedbackReason('other', details)
   )();
 
 export const logExpandSmartSnippetSuggestion = (